Add new_peer function and PeersList_to_string function***
This commit adds a new function called new_peer to the peer.c file. This function allows for the creation of a new peer and adds it to the list of peers. It also includes a new function called PeersList_to_string, which converts the list of peers to a string representation.